+++ src/components/mod.rs
@@ -6,6 +6,7 @@ pub mod debug;
 pub mod linescore;
 pub mod live_game;
 pub mod matchup;
+pub mod pitch_event;
 pub mod pitches;
 pub mod plays;
 pub mod schedule;
