+++ src/components/pitches.rs
@@ -1,7 +1,6 @@
+use crate::components::pitch_event::PitchEvent;
 use crate::components::strikezone::{DEFAULT_SZ_BOT, DEFAULT_SZ_TOP};
 use crate::components::util::convert_color;
-use crate::ui::plays::{BLUE, SCORING_SYMBOL};
-
 use mlb_api::live::LiveResponse;
 use mlb_api::plays::PlayEvent;
 use tui::{
@@ -15,25 +14,7 @@ pub struct Pitches {
     pub pitch_events: Vec<PitchEvent>,
 }
 
-#[derive(Debug, Eq, PartialEq, Clone, Copy)]
-pub enum PitchEventType {
-    Pitch,
-    Running,
-    /// pickoff attempt, mound visit, pinch hitter, etc
-    Other,
-}
-
 #[derive(Debug)]
-pub struct PitchEvent {
-    pub event_type: PitchEventType,
-    pub description: String,
-    pub pitch: Option<Pitch>,
-    pub is_scoring: Option<bool>,
-    pub away_score: Option<u8>,
-    pub home_score: Option<u8>,
-}
-
-#[derive(Clone, Debug)]
 pub struct Pitch {
     #[allow(dead_code)]
     pub strike: bool,
@@ -76,55 +57,6 @@ impl From<mlb_api::plays::Count> for Count {
         Self {
             balls: value.balls,
             strikes: value.strikes,
-        }
-    }
-}
-
-impl From<&PlayEvent> for PitchEvent {
-    fn from(play: &PlayEvent) -> Self {
-        let pitch = match play.is_pitch {
-            true => Some(Pitch::from(play)),
-            false => None,
-        };
-        let event_type = match (play.is_pitch, play.is_base_running_play) {
-            (true, _) => PitchEventType::Pitch,
-            (false, Some(true)) => PitchEventType::Running,
-            (false, _) => PitchEventType::Other,
-        };
-        Self {
-            event_type,
-            description: play.details.description.clone().unwrap_or_default(),
-            pitch,
-            is_scoring: play.details.is_scoring_play,
-            away_score: play.details.away_score,
-            home_score: play.details.home_score,
-        }
-    }
-}
-
-impl PitchEvent {
-    /// Convert a pitch event into a TUI Line item.
-    /// If it's a pitch, display the pitch information.
-    /// Otherwise, display the description.
-    pub fn as_lines(&self, debug: bool) -> Option<Vec<Line>> {
-        if self.event_type == PitchEventType::Pitch && self.pitch.is_some() {
-            Some(self.pitch.as_ref().unwrap().as_lines(debug))
-        } else if self.event_type != PitchEventType::Pitch {
-            let mut spans = Vec::new();
-            if self.is_scoring.unwrap_or(false) {
-                spans.push(Span::styled(
-                    format!(" {SCORING_SYMBOL}"),
-                    Style::default().fg(BLUE),
-                ));
-                if let (Some(away), Some(home)) = (self.away_score, self.home_score) {
-                    spans.push(Span::raw(format!(" {away}-{home}")));
-                }
-            }
-            spans.push(Span::raw(format!(" {}", self.description)));
-            spans.push(Span::raw(" "));
-            Some(vec![Line::from(spans)])
-        } else {
-            None
         }
     }
 }
